Converted save system
+ Converted headbutt tree system.
+ Converted palette maps for tilesets.
* Fixed trainer palette loading issue.
* Experimental tiled map loading extension (incomplete!)
* Fixed const correctness for PrintNum.

+++ data/tilesets.h
@@ -5,7 +5,7 @@ struct Tileset {
     const char* const metaPath;
     const uint8_t* const coll;
     const struct TilesetAnim* const anim;
-    const uint16_t* const palMap;
+    const uint8_t* const palMap;
 };
 
 extern const struct Tileset Tilesets[];
